students/k33421/Novikov_Alexey/lr2/requirements.txt
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.29
psycopg[c]==3.1.18
httpx==0.27.0